Remake StandardColorPicker using new controls
