+++ src/ColorPicker/ColorDisplay.xaml
@@ -1,4 +1,4 @@
-﻿<local:PickerControlBase x:Class="ColorPicker.ColorDisplay"
+﻿<local:DualPickerControlBase x:Class="ColorPicker.ColorDisplay"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -70,4 +70,4 @@
             </Grid>
         </Grid>
     </Grid>
-</local:PickerControlBase>
+</local:DualPickerControlBase>
