+++ src/ColorPicker/ColorDisplay.xaml.cs
@@ -1,29 +1,12 @@
-﻿using ColorPicker.Models;
-using System.Windows;
+﻿using System.Windows;
 
 namespace ColorPicker
 {
-    public partial class ColorDisplay : PickerControlBase
+    public partial class ColorDisplay : DualPickerControlBase
     {
-        public static readonly DependencyProperty SecondColorStateProperty =
-            DependencyProperty.Register(nameof(SecondColorState), typeof(ColorState), typeof(ColorDisplay),
-                new PropertyMetadata(new ColorState(1, 1, 1, 1, 0, 0, 1), OnSecondColorStatePropertyChange));
-        private SecondColorDecorator secondColorDecorator;
-        public ColorState SecondColorState
-        {
-            get => (ColorState)GetValue(SecondColorStateProperty);
-            set => SetValue(SecondColorStateProperty, value);
-        }
 
-        public NotifyableColor SecondColor
-        {
-            get;
-            set;
-        }
         public ColorDisplay() : base()
         {
-            secondColorDecorator = new SecondColorDecorator(this);
-            SecondColor = new NotifyableColor(secondColorDecorator);
             InitializeComponent();
         }
 
@@ -32,10 +15,6 @@ namespace ColorPicker
             var temp = ColorState;
             ColorState = SecondColorState;
             SecondColorState = temp;
-        }
-        private static void OnSecondColorStatePropertyChange(DependencyObject d, DependencyPropertyChangedEventArgs args)
-        {
-            ((ColorDisplay)d).SecondColor.UpdateEverything((ColorState)args.OldValue);
         }
     }
 }
