+++ src/ColorPicker/Models/ISecondColorStorage.cs
@@ -1,0 +1,7 @@
+﻿namespace ColorPicker.Models
+{
+    interface ISecondColorStorage
+    {
+        public ColorState SecondColorState { get; set; }
+    }
+}
