+++ src/ColorPicker/PortableColorPicker.xaml
@@ -32,20 +32,24 @@
                     </ImageBrush.RelativeTransform>
                 </ImageBrush>
             </Border.Background>
-            <ToggleButton Style="{StaticResource ColorToggleButtonStyle}" Padding="0" Margin="0" Height="20" Width="40"
+            <!--<ToggleButton Style="{StaticResource ColorToggleButtonStyle}" Padding="0" Margin="0" Height="20" Width="40"
                           IsEnabled="{Binding Path=IsOpen, ElementName=popup, Converter={StaticResource BoolToInvertedBoolConverter}}"
                           x:Name="toggleButton" BorderThickness="0" Panel.ZIndex="1"
-                          Background="{Binding Path=SelectedColor, ElementName=colorPicker, Converter={StaticResource ColorToBrushConverter}}">
+                          -->
+            <!--Background="{Binding Path=SelectedColor, ElementName=colorPicker, Converter={StaticResource ColorToBrushConverter}}"-->
+            <!-->
                 <Border Height="20" Width="40" BorderThickness="0" BorderBrush="Transparent" Background="Transparent" />
-            </ToggleButton>
+            </ToggleButton>-->
         </Border>
         <Popup Name="popup" StaysOpen="False" IsOpen="{Binding Path=IsChecked, ElementName=toggleButton, Mode=TwoWay}">
             <Border BorderThickness="1" BorderBrush="Black" Background="White">
-                <local:StandardColorPicker x:Name="colorPicker"
-                                   SelectedColor="{Binding ElementName=uc, Path=SelectedColor, Mode=TwoWay}"
+                <!--<local:StandardColorPicker x:Name="colorPicker"
+                                   -->
+                <!--SelectedColor="{Binding ElementName=uc, Path=SelectedColor, Mode=TwoWay}"
                                    SecondaryColor="{Binding ElementName=uc, Path=SecondaryColor, Mode=TwoWay}" 
-                                   HueSmallChange="{Binding ElementName=uc, Path=HueSmallChange}"
-                                   SmallChange="{Binding ElementName=uc, Path=SmallChange}"/>
+                                   HueSmallChange="{Binding ElementName=uc, Path=HueSmallChange}"-->
+                <!--
+                                   SmallChange="{Binding ElementName=uc, Path=SmallChange}"/>-->
             </Border>
         </Popup>
     </Grid>
